Add files via upload
Revenue insights in Hospitality Domain - Power BI
A Hotel chain faced loss in their market share and revenue over a few months. To understand the cause of the loss, they need to analyze this. I created a dashboard using 3 months data.
This helped the hotel chain to gain 20% of their revenue and market shares by the next month.